auth-service/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
pydantic[email]==2.5.0

